Add files via upload
Blockchain is the technology which allows cryptocurrencies to exist. The purpose of this document is to consider the future prospects of blockchain technology and cryptocurrencies in financial markets, as well as in others. Pakistan’s financial industry should enable availability of easy to understand educational material for potential investors and provide cheap investment advisory to people seeking to invest in the capital markets. However, a majority of middle class and lower-class investors shy away from investments in financial markets assuming illegitimacy. Financial institutions were the first to pay notice to it, as it was in simple words a new payment system.